🔥 Remove old MonacoCodeBlock
